ai-service/requirements.txt
# AI 服務依賴
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# OCR 和圖像處理
pytesseract==0.3.10
easyocr==1.7.0
google-cloud-vision==3.4.4
google-cloud-translate==3.11.1

# 自然語言處理
transformers==4.36.2
spacy==3.7.2
jieba==0.42.1

# 數據處理和分析
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0

# 機器學習和深度學習
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2

# 工具庫
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
celery==5.3.4

# 開發和測試
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
